Changed Suits to scriptable objects and added example item
