+++ Assets/Prefabs/Gameplay.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 185b7363bb21ad7409bac02dadb89f7e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
