+++ Assets/Prefabs/Gameplay/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 318f9929a1b914a4fb86e95c47de39d1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
