+++ Assets/Prefabs/Gameplay/Items/DummyItem.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 74f74aa56bb09ca499e3ae4f8239f6bc
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
