+++ Assets/ScriptableObjects/PlayerSuits.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 603c0cac55f95be4498ebae7dc04ffa4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
