+++ Assets/ScriptableObjects/PlayerSuits/DummySuit.asset
@@ -1,0 +1,21 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8c15c02802f1830488d747b8cf863763, type: 3}
+  m_Name: DummySuit
+  m_EditorClassIdentifier: 
+  MainItem: {fileID: 6857444656478776864, guid: 74f74aa56bb09ca499e3ae4f8239f6bc,
+    type: 3}
+  SideItem: {fileID: 6857444656478776864, guid: 74f74aa56bb09ca499e3ae4f8239f6bc,
+    type: 3}
+  m_defaultMaxHealth: 15
+  m_defaultWalkingVelocity: 8
+  m_defaultRunningMultiplier: 1.5
