+++ Assets/ScriptableObjects/PlayerSuits/DummySuit.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee926e424cbe13847b4b097df7dbec2c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
