+++ Assets/Scripts/Common/Gameplay/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d3bfd1292cc5a5418afbf0ff688bef7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
