+++ Assets/Scripts/Common/Gameplay/Items/ExampleItem.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+using System.Collections;
+
+namespace ubv.common.gameplay
+{
+    public class ExampleItem : PlayerItem
+    {
+        protected override void ItemActivation()
+        {
+            Debug.Log("Dummy item activated");
+        }
+    }
+}
