+++ Assets/Scripts/Common/Gameplay/Items/ExampleItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 698470eab5972114dbdc62e3f4b01647
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
