+++ Assets/Scripts/Common/Gameplay/PlayerController.cs
@@ -18,21 +18,27 @@ namespace ubv.common.gameplay
             m_defaultStats = new PlayerStats(m_defaultMaxHealth, m_defaultWalkingVelocity, m_defaultRunningMultiplier);
         }
 
-        // Use this for initialization
-        void Start()
+        public void EquipSuit(PlayerSuit suit)
         {
-
-        }
-
-        // Update is called once per frame
-        void Update()
-        {
-
+            m_currentSuit = suit;
+            m_currentSuit.Init();
+            m_currentSuit.MainItem = GameObject.Instantiate(m_currentSuit.MainItem, transform).GetComponent<PlayerItem>();
+            m_currentSuit.SideItem = GameObject.Instantiate(m_currentSuit.SideItem, transform).GetComponent<PlayerItem>();
         }
 
         public PlayerStats GetStats()
         {
             return m_currentSuit ? m_currentSuit.Stats : m_defaultStats;
+        }
+        
+        public void ActivateMainItem()
+        {
+            m_currentSuit?.MainItem.Activate();
+        }
+
+        public void ActivateSideItem()
+        {
+            m_currentSuit?.SideItem.Activate();
         }
     }
 }
